TEST: Extend UCP tag tests, and run with random message sizes.
